feature: code braket.circuits.unitary_calculation.as_unitary()
