+++ src/braket/circuits/circuit.py
@@ -877,85 +877,6 @@ the number of qubits in target_qubits must be the same as defined by the multi-q
             basis_rotation_instructions=ir_basis_rotation_instructions,
         )
 
-    @staticmethod
-    def _einsum_subscripts(targets: QubitSet, qubit_count: int) -> str:
-        target_count = len(targets)
-
-        gate_left_indexes = list(range(target_count))
-        un_left_indexes = list(range(target_count, target_count + qubit_count))
-        un_right_indexes = list(range(target_count + qubit_count, target_count + 2 * qubit_count))
-
-        gate_right_indexes = [un_left_indexes[-1 - target] for target in targets]
-
-        result_left_indexes = un_left_indexes.copy()
-        for pos, target in enumerate(targets):
-            result_left_indexes[-1 - target] = gate_left_indexes[pos]
-
-        return (
-            gate_left_indexes + gate_right_indexes,
-            un_left_indexes + un_right_indexes,
-            result_left_indexes + un_right_indexes,
-        )
-
-    def as_unitary(self) -> np.ndarray:
-        """
-        Returns the unitary matrix representation of the entire circuit.
-        *Note*: The performance of this method degrades with qubit count. It might be slow for
-        qubit count > 10.
-
-        Returns:
-            np.ndarray: A numpy array with shape (2^qubit_count, 2^qubit_count) representing the
-                circuit as a unitary. *Note*: For an empty circuit, an empty numpy array is
-                returned (`array([], dtype=complex128)`)
-
-        Raises:
-            TypeError: If circuit is not composed only of `Gate` instances,
-                i.e. a circuit with `Noise` operators will raise this error.
-
-        Examples:
-            >>> circ = Circuit().h(0).cnot(0, 1)
-            >>> circ.as_unitary()
-            array([[ 0.70710678+0.j,  0.70710678+0.j,  0.        +0.j,
-                     0.        +0.j],
-                   [ 0.        +0.j,  0.        +0.j,  0.70710678+0.j,
-                    -0.70710678+0.j],
-                   [ 0.        +0.j,  0.        +0.j,  0.70710678+0.j,
-                     0.70710678+0.j],
-                   [ 0.70710678+0.j, -0.70710678+0.j,  0.        +0.j,
-                     0.        +0.j]])
-        """
-        qubits = self.qubits
-        if not qubits:
-            return np.zeros(0, dtype=complex)
-        qubit_count = max(qubits) + 1
-
-        unitary = np.eye(2 ** qubit_count, dtype=complex)
-        un_tensor = np.reshape(unitary, qubit_count * [2, 2])
-
-        for instr in self.instructions:
-            if not isinstance(instr.operator, Gate):
-                raise TypeError("Only Gate operators are supported to build the unitary")
-
-            matrix = instr.operator.to_matrix()
-            targets = instr.target
-
-            gate_indexes, un_indexes, result_indexes = Circuit._einsum_subscripts(
-                targets, qubit_count
-            )
-            gate_matrix = np.reshape(matrix, len(targets) * [2, 2])
-
-            un_tensor = np.einsum(
-                gate_matrix,
-                gate_indexes,
-                un_tensor,
-                un_indexes,
-                result_indexes,
-                dtype=complex,
-                casting="no",
-            )
-
-        return np.reshape(un_tensor, 2 * [2 ** qubit_count])
-
     def _copy(self) -> Circuit:
         copy = Circuit().add(self.instructions)
         copy.add(self.result_types)
